# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DuckLoad API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/students/onboarding/address/{studentRecordID}": {
            "put": {
                "description": "Updates the address details (region, province, city, barangay) for a student.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Save Address Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student Record ID",
                        "name": "studentRecordID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Address Info Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/students.UpdateAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message: Address information saved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/students/onboarding/base": {
            "post": {
                "description": "Creates a new student record or updates the basic profile information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Create or Update Base Profile",
                "parameters": [
                    {
                        "description": "Base Profile Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/students.CreateStudentRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns {student_record_id: \u003cid\u003e}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/students/onboarding/education/{studentRecordID}": {
            "put": {
                "description": "Updates the educational background information for a student.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Save Education Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student Record ID",
                        "name": "studentRecordID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Education Info Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/students.UpdateEducationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message: Education information saved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/students/onboarding/family/{studentRecordID}": {
            "put": {
                "description": "Updates the family information (including parents and guardians) for a student.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Save Family Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student Record ID",
                        "name": "studentRecordID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Family Info Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/students.UpdateFamilyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message: Family information saved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/students/profile/address/{studentRecordID}": {
            "get": {
                "description": "Retrieves the residential address details for the student.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Get Address Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student Record ID",
                        "name": "studentRecordID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/students.StudentAddress"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid student record ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get address info",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/students/profile/base/{userID}": {
            "get": {
                "description": "Retrieves the basic profile information for a student using their User ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Get Student Base Profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/students.StudentRecord"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get base profile",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/students/profile/education/{studentRecordID}": {
            "get": {
                "description": "Retrieves the student's educational background and history.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Get Education Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student Record ID",
                        "name": "studentRecordID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/students.EducationalBackground"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid student record ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get education info",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/students/profile/family/{studentRecordID}": {
            "get": {
                "description": "Retrieves family details associated with a specific student record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Get Family Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student Record ID",
                        "name": "studentRecordID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/students.FamilyBackground"
                        }
                    },
                    "400": {
                        "description": "Invalid student record ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get family info",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "students.CreateStudentRecordRequest": {
            "type": "object",
            "required": [
                "birthDate",
                "civilStatusTypeId",
                "course",
                "genderId",
                "heightCm",
                "mobileNo",
                "placeOfBirth",
                "religionTypeId",
                "studentNumber",
                "weightKg",
                "yearLevel"
            ],
            "properties": {
                "birthDate": {
                    "type": "string"
                },
                "civilStatusTypeId": {
                    "type": "integer"
                },
                "course": {
                    "type": "string"
                },
                "genderId": {
                    "type": "integer"
                },
                "goodMoralStatus": {
                    "description": "Booleans default to false if missing",
                    "type": "boolean"
                },
                "hasDerogatoryRecord": {
                    "type": "boolean"
                },
                "heightCm": {
                    "type": "number"
                },
                "mobileNo": {
                    "type": "string"
                },
                "placeOfBirth": {
                    "type": "string"
                },
                "religionTypeId": {
                    "type": "integer"
                },
                "section": {
                    "description": "Optional",
                    "type": "string"
                },
                "studentNumber": {
                    "type": "string"
                },
                "weightKg": {
                    "type": "number"
                },
                "yearLevel": {
                    "type": "integer"
                }
            }
        },
...
        "students.EducationalBackground": {
            "type": "object",
            "properties": {
                "awards": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "educationalLevelId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "schoolName": {
                    "type": "string"
                },
                "schoolType": {
                    "type": "string"
                },
                "studentRecordId": {
                    "type": "integer"
                },
                "yearCompleted": {
                    "type": "string"
                }
            }
        },
...
        "students.UpdateAddressRequest": {
            "type": "object",
            "required": [
                "addresses"
            ],
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/students.StudentAddressDTO"
                    }
                }
            }
        },
        "students.UpdateEducationRequest": {
            "type": "object",
            "required": [
                "educationalBackgrounds"
            ],
            "properties": {
                "educationalBackgrounds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/students.EducationalBGDTO"
                    }
                }
            }
        },
...
    }
}